Set exports field
Without this the default entry point is index.js, which we don't have.

Publishing raw typescript is currently discouraged despite
type-stripping being a thing for all supported Node versions. But doing
transpilation here probably means: setting up esbuild and
unplugin-isolated-decl and adding a build script because esbuild doesn't
allow specifying plugins in the CLI and then running the build script in
CI

which isn't much, but we can also just add one line to set the entry
point.

There are two fields for setting the entry point (what
requiring/importing the package name resolves to):

- "main", taking one file; importing other files in the folder is
  allowed
- "exports", taking one file or an object containing multiple entry
  points; only the listed entry points can be imported

I'm choosing the latter because of the recommended reason.
https://nodejs.org/api/packages.html#package-entry-points