switch back to deno
This allows us to use URL imports, to grab zx in testing for instance.

t.step actually works as a way to write subtests despite
/runtime/fundamentals/testing/ making it sound like it's (only) for
setup and teardown.

(describe from @std/testing/bdd does not support an async body, only
test/it does, which is a very annoying limitation.)

We also use @std/assert instead of @std/expect because the latter uses a
slightly older version of @std/internal making the package graph more of
a mess